ingeleverde versie

+++ task7/6-contest/leaderboard.py
@@ -31,11 +31,11 @@ def getGroupName(file):
 def gamesStats(games):
    scores = [game.state.getScore() for game in games]
    timeouts = [game.state.getTimeout() for game in games]
-   moves  = [game.numMoves for game in games]
+   moves  = [max(1,game.numMoves) for game in games]
    times  = [games[i].totalAgentTimes[0]/float(moves[i]) for i in range(len(games))]
    wins = [game.state.isWin() for game in games]
-   winRate = wins.count(True)/ float(len(wins))
-   avescore = sum(scores) / float(len(scores))
+   winRate = wins.count(True)/ max(1,float(len(wins)))
+   avescore = sum(scores) / max(1,float(len(scores)))
    return (scores,timeouts,moves,times,wins,winRate,avescore)
 
 
@@ -44,13 +44,13 @@ def generateHtml(games, name=None):
       scores = []; timeouts=[]; moves=[]; times=[]; wins=[]; winRate=[]; avescore=[];
       for laygames in games: # average the per-layout averages
          (lscores,ltimeouts,lmoves,ltimes,lwins,lwinRate,lavescore)=gamesStats(laygames)
-         scores.append(   sum(lscores)  /float(len(lscores))   )
-         timeouts.append( sum(ltimeouts)/float(len(ltimeouts)) )
-         moves.append(    sum(lmoves)   /float(len(lmoves))    )
-         times.append(    sum(ltimes)   /float(len(ltimes))    )
-         wins.append(     sum(lwins)    /float(len(lwins))     )
+         scores.append(   sum(lscores)  /float(max(1,len(lscores)))   )
+         timeouts.append( sum(ltimeouts)/float(max(1,len(ltimeouts))) )
+         moves.append(    sum(lmoves)   /float(max(1,len(lmoves)))    )
+         times.append(    sum(ltimes)   /float(max(1,len(ltimes)))    )
+         wins.append(     sum(lwins)    /float(max(1,len(lwins)))     )
          avescore.append(lavescore)
-      avescore = sum(avescore) / float(len(avescore))
+      avescore = sum(avescore) / float(max(1,len(avescore)))
    else:
       (scores,timeouts,moves,times,wins,winRate,avescore)=gamesStats(games)
 
@@ -73,46 +73,66 @@ def generateHtml(games, name=None):
    return (avescore,htmlout)
 
 def runFile(file,replay=False,args=[]):
+  failed = True
+  score  = 0  
+  htmlsummary= ''
   fileoutput = ''
+  games=None
   # Fetch student number
   print(file)
   studentNumber = getGroupName(file)
   if studentNumber is None: studentNumber=''
-  studentNumber=studentNumber.replace('20',' ')
-  print('Trying ' + studentNumber) 
+  print('Trying ' + studentNumber + '\nfrom file: ' + file) 
   fileoutput += studentNumber + ':' + '\n'
   #keyboard()
 
-  
   # Move their competitionAgents.py to the environment
-  try: 
-    shutil.copy(file,solutionFile)
-  except: 
-    pass
+  if not file == solutionFile :
+    try:
+      shutil.copyfile(file,solutionFile)
+    except shutil.SameFileError as err: # samefile isn't a bug
+      pass
+    except IOError as err:
+      print('Error couldnt copy over the solution file: ' + file)
+      return (score,htmlsummary)
 
   # Run the environment
   output = ''
   failed = False
-  try:
-    print('Trying ' + studentNumber)
-    # force reload of student code -- needed even for load to get the class definitions correctly
-    imp.reload(sys.modules[solutionModule]) 
-    if os.path.isfile(studentNumber + '_games.pk') and not replay:
-      games = loadobject(studentNumber + '_games.pk')
-    else:
-      games = pacman.cmdlineRunGames(execList + args)
-      saveobject(studentNumber + '_games.pk',games) # save so don't have to re-run later
-    
-    score,htmlsummary = generateHtml(games,studentNumber)    
 
-  except subprocess.CalledProcessError as e:
-    failed = True
-    output = e.output
+  # force reload of student code -- needed even for load to get the class definitions correctly
+  try :
+    print(sys.modules[solutionModule])
+    imp.reload(sys.modules[solutionModule]) 
+  except:
+    failed=True
+    output = 'Error could not load solution module';
+    print(output)
+    
+  if not failed :
+      if os.path.isfile(studentNumber + '_games.pk') and not replay:
+        games = loadobject(studentNumber + '_games.pk')
+      else:
+        try:
+          nargs = execList;
+          print('Running with arguments :' + '  '.join(execList + args) )
+          games = pacman.cmdlineRunGames(execList + args)
+        except subprocess.CalledProcessError as e:          
+          failed = True
+          output = e.output
+      if not failed:
+        try:
+          saveobject(studentNumber + '_games.pk',games) # save so don't have to re-run later
+        except:
+          print('Couldnt pickel the saved games...')
+        score,htmlsummary = generateHtml(games,studentNumber)    
+
   # Print result
   if failed:
     fileoutput += 'Failed :\n' + output + '\n\n'
   else:
     fileoutput += 'Output :\n' + output + '\n\n'
+  print(fileoutput)
   return (score,htmlsummary)
 
 def help():
@@ -125,7 +145,7 @@ If no studentNumber argument is given, it will run all pairs of search.py and se
 
 -h|--help\t\tDisplays this help text.
 --dir=[directory]\tSets the default directory with student solutions to the entered one. By default it is `{defaultDir}`.
---num=[studentNumber]\tScript will only execute search.py and searchAgents.py from this student.
+--num=[studentNumber]\tScript will only execute search.py and searchAgents.py from this student group.
 
 Examples:
 ./{filename}\t\t\tRuns all file pairs in {defaultDir}/
@@ -136,15 +156,22 @@ Examples:
 def main(argv):
   defaultDirectory = 'studentsolutions'
   studentNumber = None
+  replay=False
+  args=None
+
+  # Remove the old fileoutput file
+  if os.path.isfile(logfileName):
+     shutil.copy(logfileName,logfileName + '.bak')
+     os.remove(logfileName)
   logfile = open(logfileName,'w');
 
   # Parse the arguments
+  opts=[]
   try:
      opts, args = getopt.getopt(argv, 'h', ['help', 'dir=', 'num=', 'replay'])
   except getopt.GetoptError:
-    opts=[]
     _, e, _ = sys.exc_info()
-    if e.msg.find('not recognized')<0: # pass through unrecog optoins
+    if e.msg.find('not recognized')<0: # pass through unrecog options
        print((help()))
        sys.exit(2)
 
@@ -157,15 +184,9 @@ def main(argv):
     elif opt == '--num':
        studentNumber = arg
     elif opt == '--replay':
-       replay = True
-       
-  # Remove the old fileoutput file
-  if os.path.isfile('output.txt'):
-     try:
-       os.remove('output.txt')
-     except:
-       pass
-   
+       replay = True       
+
+  print(args)
   fileoutput = ''
 
   if not os.path.isdir(defaultDirectory):
@@ -179,29 +200,35 @@ def main(argv):
     # extract student number and remove duplicates
     oldgroup=None
     files=[]
-    #print '\n'.join(solnfiles)
-    for file in solnfiles:
-      groupName = getGroupName(file)
-      if not groupName == None :
-        if groupName == oldgroup:
-          files[-1]=file
-        else:
-          oldgroup = groupName
-          files.append(file)
-    #print '\n'.join(files)
+    #print('\n'.join(solnfiles))
+    if len(solnfiles)==1:
+      files=solnfiles
+    else:
+      for file in solnfiles:
+        groupName = getGroupName(file)
+        if not groupName == None :
+          if groupName == oldgroup:
+            files[-1]=file
+          else:
+            oldgroup = groupName
+            files.append(file)
+    print('\n'.join(files))
     for file in files:
       score,output = runFile(file,replay,args)
+      print(output)
       logfile.write(output)
       info.append((score,output))
   else:
     # If you want to run one at the time using the student number
     for file in glob.glob(defaultDirectory + '/*' + solutionFile):
       groupName = getGroupName(file)
-      if studentNumber == groupName or ( groupName is None and studentNumber=='' ):
+      if 'Groep ' + studentNumber == groupName or ( groupName is None and studentNumber=='' ):
         score,output = runFile(file,replay,args)
-    logfile.write(output)
-    info.append((score,output))
+        print(output)
+        logfile.write(output)
+        info.append((score,output))
 
+  logfile.close()
   # sort the results based on score
   info = sorted(info,reverse=True)
   shutil.copy(htmlstyle,htmloutputfile) # copy in style info
@@ -215,25 +242,31 @@ def main(argv):
      hfile.write(output)
   # write the footer
   hfile.write("</table>");
+  hfile.close()
   return info
 
 if __name__ == "__main__":
-  # make back up of current solution file, if it's not already there!
+  # make back up of current solution file
+  if os.path.isfile(solutionFile):
+    shutil.copy(solutionFile,solutionFile+ ".bak");
   if not os.path.isfile(solutionFile+".orig"):
     shutil.copy(solutionFile,solutionFile + ".orig")  
   try:
     __import__(solutionModule)
-  except: # current solutinFile is broken, copy back the .orig 
-    print("Warning: " + solutionFile + " is broken!  reverting to " + solutionFile + ".orig")
-    print("     your broken one is saved to : " + solutionFile + ".bak")
-    shutil.copy(solutionFile,solutionFile + ".bak") # restore backup    
-    shutil.copy(solutionFile + ".orig",solutionFile) # restore backup
-    __import__(solutionModule)
-    
+  except :
+    print("Couldnt import the file " + solutionFile)
+    print("copy back from orginal and re-try:" + solutionFile + ".orig")
+    sys.exit(-1)
+      
   # ask for arguments if none given
   if len(sys.argv) == 1:
     if sys.version_info < (3,0):
+      print("Python version <3.0 detected, YMMV")
       sys.argv.extend([x for x in re.split(r' *',raw_input("Enter any command line arguments?")) if x!=''])
     else:
       sys.argv.extend([x for x in re.split(r' *',input("Enter any command line arguments?")) if x!=''])
   main(sys.argv[1:])
+
+  # copy the orginal file back
+  #if os.path.isfile(solutionFile + '.orig'):
+  #  shutil.copy(solutionFile + ".orig",solutionFile) # restore backup
